+++ trans_sim/artificial_translation.py
@@ -5,15 +5,14 @@ import argparse
 import numpy as np
 from pydicom import dcmread
 import subprocess
-import sys
 
 def apply_translation(args, dir, extension):
     """
     Resample volume, creating a copy of the reference volume translated in a sinusoidal fashion
     """
-
-    os.remove('translation.csv')
-    f = open('translation.csv', 'w') # reopen in append mode
+    if os.path.exists('parameters.csv'):   
+        os.remove('parameters.csv')
+    f = open('parameters.csv', 'w') # reopen in append mode
 
     try:
         reference = sitk.ReadImage(args.inVol) #read in vol
@@ -27,13 +26,10 @@ def apply_translation(args, dir, extension):
     reader.LoadPrivateTagsOn()
     reader.ReadImageInformation()
 
-    #get file extension for saving purporses
-    extension = os.path.splitext(args.inVol)
-
     for i in range(int(args.num_dicoms)):
         #set translation array
         sin = math.sin(float(args.period)*i)
-        translation_array = (((args.x) * sin), (args.y * sin), (args.z * sin))
+        translation_array = ((args.x * sin), (args.y * sin), (args.z * sin))
         transform = sitk.AffineTransform(3)
         transform.SetTranslation(translation_array)
 
@@ -54,10 +50,10 @@ def apply_translation(args, dir, extension):
             transformed_image.SetMetaData(j, reader.GetMetaData(j))
 
         sitk.WriteImage(transformed_image, os.path.join(dir, f'translated_{str(i).zfill(4)}{extension[1]}'))  #write slice to directory
-        print(f'Volume Translation {i}, Translation: {translation_array}')
+        print(f'Volume {i} Translation: Translation X: {translation_array[0]} mm, Translation Y: {translation_array[1]} mm, Translation Z: {translation_array} mm')
 
         # plot simulated motion transformations for a reference plot, only if the flag is used
-        if args.refplot: write_simulated_data(f, args, sin, i)
+        write_simulated_data(f, args, translation_array, i)
 
     f.close()
 
@@ -96,8 +92,7 @@ def vvr(refVol, voldir):
     dirmapping = os.getcwd() + ":" + "/data"
     dockerprefix = ["docker","run","--rm", "-it", "--init", "-v", dirmapping,
         "--user", str(os.getuid())+":"+str(os.getgid())]
-    print(dockerprefix)
-
+    
     inputTransformFileName = "identity-centered.tfm"
     subprocess.run( dockerprefix +  [ "crl/sms-mi-reg", "crl-identity-transform-at-volume-center.py", 
     "--refvolume", refVol, 
@@ -108,45 +103,64 @@ def vvr(refVol, voldir):
         vol = os.path.join(voldir, volname)
         outTransFile = str(i).zfill(4)
 
-        subprocess.run( dockerprefix + ["crl/sms-mi-reg", "sms-mi-reg", 
-        refVol,
-        inputTransformFileName,
-        outTransFile,
-        vol] )
+        if i == 0: #use identity-centered.tfm for only the first registration
+            subprocess.run( dockerprefix + ["crl/sms-mi-reg", "sms-mi-reg", 
+            refVol,
+            inputTransformFileName,
+            outTransFile,
+            vol] )
+        else: #use the most recent written transform file for the next registration
+            inputTransformFileName = f"sliceTransform{str(i - 1).zfill(4)}.tfm"
 
-def write_simulated_data(f, args, sin, i):
+            print( dockerprefix + ["crl/sms-mi-reg", "sms-mi-reg", 
+            refVol,
+            inputTransformFileName,
+            outTransFile,
+            vol] )
+
+            subprocess.run( dockerprefix + ["crl/sms-mi-reg", "sms-mi-reg", 
+            refVol,
+            inputTransformFileName,
+            outTransFile,
+            vol] )
+
+def write_simulated_data(f, args, translation_array, i):
         #write to csv file the 6 parameters
-        transform_array =(0,0,0,(args.x * sin), (args.y * sin), (args.z * sin))
+        transform_array =(0,0,0) + translation_array
         a = np.asarray(transform_array).reshape(1,-1)
         np.savetxt(f, a, delimiter=",",fmt="%.8f")
-        transform1 = sitk.VersorRigid3DTransform()
-        transform1.SetCenter((0,0,0)) #center of rotation #COME BACK TO THIS
-        transform1.SetParameters(transform_array)
-        sitk.WriteTransform(transform1, f'./{str(i).zfill(4)}.tfm')
+        
+        if args.refplot:
+            transform1 = sitk.VersorRigid3DTransform()
+            transform1.SetCenter((0,0,0)) #center of rotation #COME BACK TO THIS
+            transform1.SetParameters(transform_array)
+            sitk.WriteTransform(transform1, f'./{str(i).zfill(4)}.tfm')
 
-        if i == (args.num_dicoms - 1):
-            dirmapping_a = os.getcwd() + ":" + "/data"
-            dockerprefix_a = ["docker","run","--rm", "-it", "--init", "-v", dirmapping_a, "--user", str(os.getuid())+":"+str(os.getgid())]
+            if i == (args.num_dicoms - 1):
+                dirmapping_a = os.getcwd() + ":" + "/data"
+                dockerprefix_a = ["docker","run","--rm", "-it", "--init", "-v", dirmapping_a, "--user", str(os.getuid())+":"+str(os.getgid())]
 
-            subprocess.run( dockerprefix_a + ["jauger/motion-monitor:latest", "0000.tfm"])
-            # Loop through files in the directory
-            for filename in os.listdir(os.getcwd()):
-                if filename.endswith(".tfm"):
-                    os.remove(filename)
+                subprocess.run( dockerprefix_a + ["jauger/motion-monitor:latest", "0000.tfm"])
+                # Loop through files in the directory
+                for filename in os.listdir(os.getcwd()):
+                    if filename.endswith(".tfm"):
+                        os.remove(filename)
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
 
     parser.add_argument('-inVol', default='./input/adultjosh.dcm', help='file path of input volume to perform artificial translation') #change this to inVol instead of -inVol. same for x y z vals
-    parser.add_argument('-x', default=2.5, help='x-axis translation sin wave magnitude. default is 0. Number can be integer or float')
-    parser.add_argument('-y', default=0, help='y-axis translation sin wave magnitude. default is 0. Number can be integer or float')
-    parser.add_argument('-z', default=0, help='z-axis translation sin wave magnitude. default is 0. Number can be integer or float')
-    parser.add_argument('-period', default=0.2, help='period of sinusoidal motion, default is 0.2. Number can be integer or float') 
+
+    parser.add_argument('-x', default=2.5, help='x-axis translation sin wave magnitude. default is 0. Number can be integer or decimal')
+    parser.add_argument('-y', default=0, help='y-axis translation sin wave magnitude. default is 0. Number can be integer or fldecimaloat')
+    parser.add_argument('-z', default=0, help='z-axis translation sin wave magnitude. default is 0. Number can be integer or decimal')
+    parser.add_argument('-period', default=0.2, help='period of sinusoidal motion, default is 0.2. Number can be integer or decimal') 
     parser.add_argument('-num_dicoms', default=40, help='number of output dicoms, default is 40. Number must be integer')
 
     parser.add_argument('--vvr', action='store_true', help='flag for performing volume to volume registration')
     parser.add_argument('--refplot', action='store_true', help='flag to create motion monitor plots for simulated motion parameters')
+    parser.add_argument('--slimm', action='store_true', help='flag to perform metadata regeneration on dicom images to use on slimm')
 
     args = parser.parse_args()
 
@@ -162,7 +176,7 @@ if __name__ == '__main__':
     apply_translation(args, directory, extension)
 
     #only do metadata regeneration step if the file is a dicom, so that it can be converted to an mrd for slimm
-    if extension[1] == '.dcm':
+    if extension[1] == '.dcm' and args.slimm:
         metadata(args.inVol, directory)
 
     #perform vvr and motion monitor if the flag is used
